+++ Color/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.2
 
 * Fix `YCbCr` conversion to derived RGB color spaces
+* Fix conversion of `toWord64 (1 :: Double)` and `toWord632 (1 :: Float)`.
 
 ## 0.1.1